admin auth+productRating

+++ BridalOrdering/Controllers/PackageController.cs
@@ -10,6 +10,7 @@ using Microsoft.AspNetCore.Routing;
 using Microsoft.Extensions.Logging;
 using  Newtonsoft.Json;
 using BridalOrdering.Models;
+using Microsoft.AspNetCore.Http;
 
 namespace BridalOrdering.Controllers
 {
@@ -27,10 +28,14 @@ namespace BridalOrdering.Controllers
             _productStore = productStore;
             
         }
+        [Authorize]
         [HttpPost]
         [Route("add")]
         public async Task<IActionResult> AddAsync([FromBody]Package model)
         {
+             var userType = User.Claims.FirstOrDefault(x => x.Type == "userType" ).Value;
+            if(userType!= UserType.ADMIN.ToString())
+                return new JsonResult(new { message = "Unauthorized" }) { StatusCode = StatusCodes.Status401Unauthorized };
             if (!ModelState.IsValid)
             {
                 return BadRequest(ModelState);
@@ -38,7 +43,6 @@ namespace BridalOrdering.Controllers
            await _store.InsertOneAsync(model);
             return Ok(CreateSuccessResponse("Created successfully"));
         }
-        [Authorize]
         [HttpGet]
         [Route("get")]
         public async Task<IActionResult> GetAllAsync()
@@ -79,11 +83,14 @@ namespace BridalOrdering.Controllers
             }
             return Ok(response);
         }
-
+        [Authorize]
         [HttpPost]
         [Route("update/{PackageId}")]
         public async Task<IActionResult> UpdateAsync([FromBody]Package model, [FromRoute] string PackageId)
         {
+             var userType = User.Claims.FirstOrDefault(x => x.Type == "userType" ).Value;
+            if(userType!= UserType.ADMIN.ToString())
+                return new JsonResult(new { message = "Unauthorized" }) { StatusCode = StatusCodes.Status401Unauthorized };
             var result = await _store.FindByIdAsync(PackageId);
             result.Name = model.Name;
             result.Image = model.Image;
@@ -92,10 +99,14 @@ namespace BridalOrdering.Controllers
             return Ok(CreateSuccessResponse("Updated successfully"));
 
         }
+        [Authorize]
         [HttpPost]
         [Route("delete/{PackageId}")]
         public async Task<IActionResult> Delete([FromRoute] string PackageId)
         {
+             var userType = User.Claims.FirstOrDefault(x => x.Type == "userType" ).Value;
+            if(userType!= UserType.ADMIN.ToString())
+                return new JsonResult(new { message = "Unauthorized" }) { StatusCode = StatusCodes.Status401Unauthorized };
            
             await _store.DeleteByIdAsync(PackageId);
             return Ok(CreateSuccessResponse("Deleted successfully"));
